feat: database manager
